move pod util functions to node/corev1/helpers
