+++ pkg/api/v1/pod/util_test.go
@@ -796,69 +796,3 @@ func TestGetContainerStatus(t *testing.T) {
 		assert.Equal(t, test.expected.status, resultStatus, "GetExistingContainerStatus: "+test.desc)
 	}
 }
-
-func TestUpdatePodCondition(t *testing.T) {
-	time := metav1.Now()
-
-	podStatus := v1.PodStatus{
-		Conditions: []v1.PodCondition{
-			{
-				Type:               v1.PodReady,
-				Status:             v1.ConditionTrue,
-				Reason:             "successfully",
-				Message:            "sync pod successfully",
-				LastProbeTime:      time,
-				LastTransitionTime: metav1.NewTime(time.Add(1000)),
-			},
-		},
-	}
-	tests := []struct {
-		status     *v1.PodStatus
-		conditions v1.PodCondition
-		expected   bool
-		desc       string
-	}{
-		{
-			status: &podStatus,
-			conditions: v1.PodCondition{
-				Type:               v1.PodReady,
-				Status:             v1.ConditionTrue,
-				Reason:             "successfully",
-				Message:            "sync pod successfully",
-				LastProbeTime:      time,
-				LastTransitionTime: metav1.NewTime(time.Add(1000))},
-			expected: false,
-			desc:     "all equal, no update",
-		},
-		{
-			status: &podStatus,
-			conditions: v1.PodCondition{
-				Type:               v1.PodScheduled,
-				Status:             v1.ConditionTrue,
-				Reason:             "successfully",
-				Message:            "sync pod successfully",
-				LastProbeTime:      time,
-				LastTransitionTime: metav1.NewTime(time.Add(1000))},
-			expected: true,
-			desc:     "not equal Type, should get updated",
-		},
-		{
-			status: &podStatus,
-			conditions: v1.PodCondition{
-				Type:               v1.PodReady,
-				Status:             v1.ConditionFalse,
-				Reason:             "successfully",
-				Message:            "sync pod successfully",
-				LastProbeTime:      time,
-				LastTransitionTime: metav1.NewTime(time.Add(1000))},
-			expected: true,
-			desc:     "not equal Status, should get updated",
-		},
-	}
-
-	for _, test := range tests {
-		resultStatus := UpdatePodCondition(test.status, &test.conditions)
-
-		assert.Equal(t, test.expected, resultStatus, test.desc)
-	}
-}
