+++ test/integration/statefulset/statefulset_test.go
@@ -32,6 +32,7 @@ import (
 	"k8s.io/client-go/dynamic"
 	clientset "k8s.io/client-go/kubernetes"
 	featuregatetesting "k8s.io/component-base/featuregate/testing"
+	corev1helpers "k8s.io/component-helpers/node/corev1"
 	apiservertesting "k8s.io/kubernetes/cmd/kube-apiserver/app/testing"
 	podutil "k8s.io/kubernetes/pkg/api/v1/pod"
 	"k8s.io/kubernetes/pkg/features"
@@ -324,7 +325,7 @@ func setPodsReadyCondition(t *testing.T, clientSet clientset.Interface, pods *v1
 				continue
 			}
 			pod.Status.Phase = v1.PodRunning
-			_, condition := podutil.GetPodCondition(&pod.Status, v1.PodReady)
+			_, condition := corev1helpers.GetPodCondition(&pod.Status, v1.PodReady)
 			if condition != nil {
 				condition.Status = conditionStatus
 				condition.LastTransitionTime = metav1.Time{Time: lastTransitionTime}
